Browser base class

Base class for browsers like Chrome, Chromium, etc. Defines the interface used by various browser implementations. It can represent a local browser process or one running remotely.

Details

The initialize() method of an implementation should set private$host and private$port. If the process is local, the initialize() method should also set private$process.

Methods

Public methods

Browser$is_local()

Browser$get_process()

Browser$get_host()

Browser$get_port()

Browser$close()

Browser$clone()

Method is_local()

Is local browser? Returns TRUE if the browser is running locally, FALSE if it's remote.

Usage

Browser$is_local()

Method get_process()

Browser process

Usage

Browser$get_process()

Method get_host()

Browser Host

Usage

Browser$get_host()

Method get_port()

Browser port

Usage

Browser$get_port()

Method close()

Close the browser

Usage

Browser$close()

Method clone()

The objects of this class are cloneable with this method.

Usage

Browser$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.