ChromoteSession class

ChromoteSession class

ChromoteSession class

Methods

Public methods

ChromoteSession$new()

ChromoteSession$init_promise()

ChromoteSession$close()

ChromoteSession$view()

ChromoteSession$is_active()

ChromoteSession$new_session()

ChromoteSession$send_command()

ChromoteSession$get_parent()

ChromoteSession$get_session_id()

ChromoteSession$get_child_loop()

ChromoteSession$wait_for()

ChromoteSession$get_auto_events()

ChromoteSession$debug_log()

ChromoteSession$invoke_event_callbacks()

ChromoteSession$mark_closed()

ChromoteSession$screenshot()

ChromoteSession$screenshot_pdf()

Method new()

Create a new ChromoteSession object.

Usage

ChromoteSession$new(
  parent = default_chromote_object(),
  width = 1200,
  height = 1600,
  targetId = NULL,
  wait_ = TRUE,
  auto_events = NULL
)

Arguments

parent

Chromote object to use; defaults to default_chromote_object()

width

Width, in pixels, of the Target to create if targetId is NULL

height

Height, in pixels, of the Target to create if targetId is NULL

targetId

Target ID of an existing target to attach to. When a targetId is provided, the width and height arguments are ignored. If NULL (the default) a new target is created and attached to, and the width and height arguments determine its viewport size.

wait_

If FALSE, return a promises::promise() of a new ChromoteSession object. Otherwise, block during initialization, and return a ChromoteSession object directly.

auto_events

If NULL (the default), use the auto_events setting from the parent Chromote object. If TRUE, enable automatic event enabling/disabling; if FALSE, disable automatic event enabling/disabling.

Returns

A new ChromoteSession object.

Method init_promise()

Usage

ChromoteSession$init_promise()

Method close()

Usage

ChromoteSession$close(wait_ = TRUE)

Method view()

Usage

ChromoteSession$view()

Method is_active()

Usage

ChromoteSession$is_active()

Method new_session()

Usage

ChromoteSession$new_session(
  width = 1200,
  height = 1600,
  targetId = NULL,
  wait_ = TRUE
)

Method send_command()

Usage

ChromoteSession$send_command(
  msg,
  callback = NULL,
  error = NULL,
  timeout = NULL
)

Method get_parent()

Usage

ChromoteSession$get_parent()

Method get_session_id()

Usage

ChromoteSession$get_session_id()

Method get_child_loop()

Usage

ChromoteSession$get_child_loop()

Method wait_for()

Usage

ChromoteSession$wait_for(p)

Method get_auto_events()

Usage

ChromoteSession$get_auto_events()

Method debug_log()

Usage

ChromoteSession$debug_log(...)

Method invoke_event_callbacks()

Usage

ChromoteSession$invoke_event_callbacks(event, params)

Method mark_closed()

Usage

ChromoteSession$mark_closed()

Method screenshot()

Usage

ChromoteSession$screenshot(
  filename = "screenshot.png",
  selector = "html",
  cliprect = NULL,
  region = c("content", "padding", "border", "margin"),
  expand = NULL,
  scale = 1,
  show = FALSE,
  delay = 0.5,
  wait_ = TRUE
)

Method screenshot_pdf()

Usage

ChromoteSession$screenshot_pdf(
  filename = "screenshot.pdf",
  pagesize = "letter",
  margins = 0.5,
  units = c("in", "cm"),
  landscape = FALSE,
  display_header_footer = FALSE,
  print_background = FALSE,
  scale = 1,
  wait_ = TRUE
)