Local Chrome process

This is a subclass of Browser that represents a local browser. It extends the Browser class with a processx::process object, which represents the browser's system process.

See also

Super class

chromote::Browser -> Chrome

Methods

Public methods

Chrome$new()

Chrome$get_path()

Chrome$clone()

Inherited methods

chromote::Browser$close()

chromote::Browser$get_host()

chromote::Browser$get_port()

chromote::Browser$get_process()

chromote::Browser$is_local()

Method new()

Create a new Chrome object.

Usage

Chrome$new(path = find_chrome(), args = get_chrome_args())

Arguments

path

Location of chrome installation

args

A character vector of command-line arguments passed when initializing Chrome. Single on-off arguments are passed as single values (e.g."--disable-gpu"), arguments with a value are given with a nested character vector (e.g. c("--force-color-profile", "srgb")). See here for a list of possible arguments. Defaults to get_chrome_args().

Returns

A new Chrome object.

Method get_path()

Browser application path

Usage

Chrome$get_path()

Method clone()

The objects of this class are cloneable with this method.

Usage

Chrome$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.