Remote Chrome process

Remote Chrome process

Remote Chrome process

Super class

chromote::Browser -> ChromeRemote

Methods

Public methods

ChromeRemote$new()

ChromeRemote$clone()

Inherited methods

chromote::Browser$close()

chromote::Browser$get_host()

chromote::Browser$get_port()

chromote::Browser$get_process()

chromote::Browser$is_local()

Method new()

Create a new ChromeRemote object.

Usage

ChromeRemote$new(host, port)

Arguments

host

A string that is a valid IPv4 or IPv6 address. "0.0.0.0" represents all IPv4 addresses and "::/0" represents all IPv6 addresses.

port

A number or integer that indicates the server port.

Method clone()

The objects of this class are cloneable with this method.

Usage

ChromeRemote$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.